+++ model/src/context/index.ts
@@ -1,3 +1,5 @@
 export * from './default-value-context';
+export * from './property-context';
+export * from './scoped-property-context';
 export * from './value-context';
 export * from './definition-context';